app: Create GetSnapshotsInfo() function
