+++ sync/sync.go
@@ -363,3 +363,87 @@ func (t *Task) getToReplica(address string) (rest.Replica, error) {
 
 	return rest.Replica{}, fmt.Errorf("Failed to find target replica to copy to")
 }
+
+const VolumeHeadName = "volume-head"
+
+func getDisks(address string) (map[string]replica.DiskInfo, error) {
+	repClient, err := replicaClient.NewReplicaClient(address)
+	if err != nil {
+		return nil, err
+	}
+
+	r, err := repClient.GetReplica()
+	if err != nil {
+		return nil, err
+	}
+
+	return r.Disks, err
+}
+
+func GetSnapshotsInfo(replicas []rest.Replica) (map[string]replica.DiskInfo, error) {
+	outputDisks := make(map[string]replica.DiskInfo)
+	for _, r := range replicas {
+		if r.Mode != "RW" {
+			continue
+		}
+
+		disks, err := getDisks(r.Address)
+		if err != nil {
+			return nil, err
+		}
+
+		for name, disk := range disks {
+			snapshot := ""
+
+			if !replica.IsHeadDisk(name) {
+				snapshot, err = replica.GetSnapshotNameFromDiskName(name)
+				if err != nil {
+					return nil, err
+				}
+			} else {
+				snapshot = VolumeHeadName
+			}
+			children := []string{}
+			for _, childDisk := range disk.Children {
+				child := ""
+				if !replica.IsHeadDisk(childDisk) {
+					child, err = replica.GetSnapshotNameFromDiskName(childDisk)
+					if err != nil {
+						return nil, err
+					}
+				} else {
+					child = VolumeHeadName
+				}
+				children = append(children, child)
+			}
+			parent := ""
+			if disk.Parent != "" {
+				parent, err = replica.GetSnapshotNameFromDiskName(disk.Parent)
+				if err != nil {
+					return nil, err
+				}
+			}
+			info := replica.DiskInfo{
+				Name:        snapshot,
+				Parent:      parent,
+				Removed:     disk.Removed,
+				UserCreated: disk.UserCreated,
+				Children:    children,
+				Created:     disk.Created,
+				Size:        disk.Size,
+			}
+			if _, exists := outputDisks[snapshot]; !exists {
+				outputDisks[snapshot] = info
+			} else {
+				// Consolidate the result of snapshot in removing process
+				if info.Removed && !outputDisks[snapshot].Removed {
+					t := outputDisks[snapshot]
+					t.Removed = true
+					outputDisks[snapshot] = t
+				}
+			}
+		}
+
+	}
+	return outputDisks, nil
+}
